refactor: use a higher-level import
