+++ ape_aws/kms/__init__.py
@@ -1,5 +1,6 @@
-from .client import KmsClient
+from .client import KmsClient, KmsKey
 
 __all__ = [
     "KmsClient",
+    "KmsKey",
 ]
